requirements.txt
Babel==2.14.0
bcrypt==4.1.2
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.1
Flask==3.0.1
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
pycparser==2.21
PyMySQL==1.1.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.3.post1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
